created separate example file + added README
